Use composer v1 to stop hack v3 tests from failing
It looks like composer recently upgraded to v2, which is now installed
by default by the composer installer script. The installer script does
allow a --version flag, but requires a specific version instead of 1.x,
so I opted for the most recent v1 version, 1.10.17.

This also removes the apt calls because the php-cli is already included in
the hhvm/hhvm docker images, and zip and unzip appeared unused. Removing
the apt commands also speeds up the test build by a good five or so
seconds.