Add test workflow for GitHub Actions
I was trying to use travis-ci.org for CI on my fork of this repo but it
was taking an hour and a half for a build to go through (and apparently
travis-ci.org is being made read-only at the end of the year in favor of
travis-ci.com). I thought I'd try making a GitHub Action to run tests
instead to see if migrating would be that hard.

This adds a `Test` workflow to run a CI matrix on almost all of the same
HHVM versions that the `travis.yml` specifies. I removed versions 3.25.3
and 3.26.3 because they were failing due to not supporting `varray`s
(this was happening on travis as well), as well as renamed `.travis.sh`
to `ci-entrypoint.sh` since the script still works totally find as an
entrypoint for the hhvm/hhvm docker image that runs the test suite.