Download upstream git for actions/checkout action
This is necessary to make sure the git command in the docker image is at least version 2.18.
Some older hhvm/hhvm:$version images have old versions of git and the actions/checkout action
downloads the repo archive over HTTP in that case. This breaks CI because the test
directory is excluded in HTTP archives in order to make the composer package size smaller by not
including unnecessary files.